requirements.txt
Django~=5.0.1
environs~=10.3.0
infisical~=1.5.0
djangorestframework~=3.14.0